+++ src/components/Redux/reducer.js
@@ -95,6 +95,7 @@ const initialState = {
   autenticado: null,
   usuario: null,
   mensaje: null,
+  cliente:false,
   /******************************* */
 
   generos: [],
@@ -169,14 +170,27 @@ function rootReducer(state = initialState, action) {
     case REGISTRO_EXITOSO:
     case LOGIN_EXITOSO:
       localStorage.setItem("token", action.payload.token);
-      localStorage.setItem("nombre", action.payload.usuario.nombre?action.payload.usuario.nombre:action.payload.usuario.nombre_cial_fantasia);
+      localStorage.setItem("nombre", action.payload.usuario?action.payload.usuario.nombre:action.payload.cliente.nombre_cial_fantasia);
+      if(action.payload.cliente){
+        return {
+          ...state,
+          autenticado: true,
+          usuario: null,
+          mensaje: null,
+          cliente:true
+        };
+      }else{
 
-      return {
-        ...state,
-        autenticado: true,
-        usuario: null,
-        mensaje: null,
-      };
+        return {
+          ...state,
+          autenticado: true,
+          usuario: null,
+          mensaje: null,
+          cliente:false
+        };
+      }
+    
+    
     case CERRAR_SESION:
     case LOGIN_ERROR:
     case REGISTRO_ERROR:
